upgrade sidebar-api & game-api, fix bugs in paged menu

+++ src/main/java/ru/stonlex/api/bukkit/game/player/GamePlayer.java
@@ -17,7 +17,7 @@ public interface GamePlayer {
     boolean isSpectator();
 
     void setSpectator();
-    void setMultiple(int multiple);
+    void setMultiple(double multiple);
     void leave();
 
 }
